clean: adapt the project to clean architecture
