+++ Cargo.toml
@@ -11,9 +11,10 @@ serde = { version = "1.0.206", features = ["derive"] }
 serde_json = "1.0.122"
 utoipa = { version = "4", features = ["rocket_extras"]}
 utoipa-swagger-ui = { version = "7", features = ["rocket"]}
-diesel = { version = "2.2.0", features = ["postgres", "chrono", "uuid"] }
+diesel = { version = "2.2.0", features = ["postgres", "chrono", "uuid", "r2d2"] }
 dotenv = "0.15.0"
 chrono = "0.4.38"
 uuid = "1.10.0"
 lettre = "=0.11.7"
 rand = "0.8.5"
+async-trait = "0.1.81"
