+++ src/adapters.rs
@@ -1,0 +1,2 @@
+pub mod controller;
+pub mod repository;
