+++ src/domain.rs
@@ -1,0 +1,3 @@
+pub mod student;
+pub mod student_pair;
+pub mod assignment;
