+++ src/domain/student.rs
@@ -1,0 +1,13 @@
+use serde::{Deserialize, Serialize};
+use utoipa::ToSchema;
+
+#[derive(Clone, Deserialize, ToSchema)]
+#[serde(rename_all = "camelCase")]
+pub struct UserInfo {
+    #[schema(example = "4622999")]
+    pub student_id: String,
+    #[schema(example = "山田")]
+    pub family_name: String,
+    #[schema(example = "太郎")]
+    pub given_name: String,
+}
