+++ src/infrastracture.rs
@@ -1,0 +1,3 @@
+pub mod router;
+pub mod models;
+pub mod schema;
