+++ src/lib.rs
@@ -1,0 +1,4 @@
+pub mod domain;
+pub mod infrastracture;
+pub mod usecase;
+pub mod adapters;
